% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite_interface.R
\name{filter_sqlite}
\alias{filter_sqlite}
\title{Send query, retrieve results and then clear result set}
\usage{
filter_sqlite(sqlite_name, sql_query, ...)
}
\arguments{
\item{sqlite_name}{A character string specifying the unquoted SQLite
database name}

\item{sql_query}{A character string containing SQL}

\item{...}{Other parameters passed on to
\code{\link[DBI]{dbGetQuery}}.}
}
\description{
Returns the result of a query as a tibble
}
