% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite_interface.R
\name{list_fields_sqlite}
\alias{list_fields_sqlite}
\title{List tables and fields names of remote SQLite tables}
\usage{
list_fields_sqlite(sqlite_name)
}
\arguments{
\item{sqlite_name}{A character string specifying the unquoted SQLite
database name}
}
\description{
Returns a list of tables and fields names of remote SQLite database
accessible through this connection from an specific path
(/R Output/SQLite Files/).
}
