% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite_interface.R
\name{list_tables_sqlite}
\alias{list_tables_sqlite}
\title{List remote SQLite tables}
\usage{
list_tables_sqlite(sqlite_name)
}
\arguments{
\item{sqlite_name}{A character string specifying the unquoted SQLite
database name}
}
\description{
Returns the unquoted names of remote SQLite tables accessible through
this connection from an specific path (/R Output/SQLite Files/).
}
