% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite_interface.R
\name{read_table_sqlite}
\alias{read_table_sqlite}
\title{Copy tibble from SQLite database}
\usage{
read_table_sqlite(sqlite_name, table_name, ...)
}
\arguments{
\item{sqlite_name}{A character string specifying the unquoted SQLite
database name}

\item{table_name}{A character string specifying the unquoted SQLite
table name}

\item{...}{Other parameters passed on to
\code{\link[DBI]{dbWriteTable}}.}
}
\description{
Reads an SQLite database table to a tibble from an
specific path (/R Output/SQLite Files/).
}
