% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary_key.R
\name{rpw_cta_cte}
\alias{rpw_cta_cte}
\title{Read, process and write a xlsx file with multiple cta_cte columns}
\usage{
rpw_cta_cte(path, write_csv = FALSE, write_sqlite = FALSE)
}
\arguments{
\item{path}{a character vector of full path name.}

\item{write_csv}{logical. Should a csv file be generated?}

\item{write_sqlite}{logical. Should a sqlite file be generated?}
}
\description{
Returns a cleaned tibble version of the mentioned xlsx file. Also, a csv
and sqlite file could be exported.
}
