% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgf.R
\name{rpw_sgf_listado_prov}
\alias{rpw_sgf_listado_prov}
\title{Read, process and write SGF's Listado de Proveedores report}
\usage{
rpw_sgf_listado_prov(path, write_csv = FALSE, write_sqlite = FALSE)
}
\arguments{
\item{path}{a character vector of full path names. If more than one path is
assigned, a dataframe that combine all of them will be return.}

\item{write_csv}{logical. Should a csv file be generated?}

\item{write_sqlite}{logical. Should a sqlite file be generated?}
}
\description{
Returns a cleaned tibble version of SGF's report. Also, a csv and sqlite
file could be exported.
}
