% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgf.R
\name{rpw_sgf_resumen_rend_obra}
\alias{rpw_sgf_resumen_rend_obra}
\title{Read, process and write SGF's Resumen de Rendiciones por obra report}
\usage{
rpw_sgf_resumen_rend_obra(
  path,
  write_csv = FALSE,
  write_sqlite = FALSE,
  overwrite_sql = FALSE
)
}
\arguments{
\item{path}{a character vector of full path names. If more than one path is
assigned, a dataframe that combine all of them will be return.}

\item{write_csv}{logical. Should a csv file be generated?}

\item{write_sqlite}{logical. Should a sqlite file be generated?}

\item{overwrite_sql}{logical. Should the sql table be overwritten? Default
set to FALSE meaning that the table would be appended.}
}
\description{
Returns a cleaned tibble version of SGF's report. Also, a csv and sqlite
file could be exported.
}
