% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siif_gastos.R
\name{rpw_siif_comprobantes_gtos_gpo_partida}
\alias{rpw_siif_comprobantes_gtos_gpo_partida}
\title{Read, process and write SIIF's gto_rpa03g report}
\usage{
rpw_siif_comprobantes_gtos_gpo_partida(
  path,
  write_csv = FALSE,
  write_sqlite = FALSE
)
}
\arguments{
\item{path}{a character vector of full path names. If more than one path is
assigned, a dataframe that combine all of them will be return.}

\item{write_csv}{logical. Should a csv file be generated?}

\item{write_sqlite}{logical. Should a sqlite file be generated?}
}
\description{
Returns a cleaned tibble version of SIIF's report. Also, a csv and sqlite
file could be exported.
}
