% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siif_contabilidad.R
\name{rpw_siif_mayor_contable}
\alias{rpw_siif_mayor_contable}
\title{Read, process and write SIIF's rcocc31 report}
\usage{
rpw_siif_mayor_contable(
  path = NULL,
  write_csv = FALSE,
  write_sqlite = FALSE,
  overwrite_sql = FALSE
)
}
\arguments{
\item{path}{a character vector of full path names. If more than one path is
assigned, a dataframe that combine all of them will be return.}

\item{write_csv}{logical. Should a csv file be generated?}

\item{write_sqlite}{logical. Should a sqlite file be generated?}

\item{overwrite_sql}{logical. Should the sql table be overwritten? Default
set to FALSE meaning that the table would be appended.}
}
\description{
Returns a cleaned tibble version of SIIF's report. Also, a csv and sqlite
file could be exported.
}
