% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siif_gastos.R
\name{rpw_siif_ppto_gtos_fte}
\alias{rpw_siif_ppto_gtos_fte}
\title{Read, process and write SIIF's rf602 report}
\usage{
rpw_siif_ppto_gtos_fte(path, write_csv = FALSE, write_sqlite = FALSE)
}
\arguments{
\item{path}{a character vector of full path names. If more than one path is
assigned, a dataframe that combine all of them will be return.}

\item{write_csv}{logical. Should a csv file be generated?}

\item{write_sqlite}{logical. Should a sqlite file be generated?}
}
\description{
Returns a cleaned tibble version of SIIF's report. Also, a csv and sqlite
file could be exported.
}
