% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siif_gastos.R
\name{rpw_siif_retenciones_por_codigo}
\alias{rpw_siif_retenciones_por_codigo}
\title{Read, process and write SIIF's rao01 report}
\usage{
rpw_siif_retenciones_por_codigo(
  path,
  write_csv = FALSE,
  write_sqlite = FALSE,
  overwrite_sql = FALSE
)
}
\arguments{
\item{path}{a character vector of full path names. If more than one path is
assigned, a dataframe that combine all of them will be return.}

\item{write_csv}{logical. Should a csv file be generated?}

\item{write_sqlite}{logical. Should a sqlite file be generated?}

\item{overwrite_sql}{logical. Should the sql table be overwritten? Default
set to FALSE meaning that the table would be appended.}
}
\description{
Returns a cleaned tibble version of SIIF's report. Also, a csv and sqlite
file could be exported.
}
