% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slave.R
\name{rpw_slave_honorarios}
\alias{rpw_slave_honorarios}
\title{Transmute old ICARO DB to new version}
\usage{
rpw_slave_honorarios(path, write_csv = FALSE, write_sqlite = FALSE)
}
\arguments{
\item{path}{a character vector of full path name of old ICARO DB.}

\item{write_csv}{logical. Should a csv file be generated?}

\item{write_sqlite}{logical. Should a sqlite file be generated?}
}
\description{
Returns a cleaned tibble version of old ICARO DB. Also, a csv and sqlite
file could be exported.
}
