% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sscc.R
\name{rpw_sscc_banco_invico}
\alias{rpw_sscc_banco_invico}
\title{Read, process and write SSCC's Movimientos Banco report}
\usage{
rpw_sscc_banco_invico(path, write_csv = FALSE, write_sqlite = FALSE)
}
\arguments{
\item{path}{a character vector of full path names. If more than one path is
assigned, a dataframe that combine all of them will be return.}

\item{write_csv}{logical. Should a csv file be generated?}

\item{write_sqlite}{logical. Should a sqlite file be generated?}
}
\description{
Returns a cleaned tibble version of SSCC's report. Also, a csv and sqlite
file could be exported.
}
