% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite_interface.R
\name{sort_table_sqlite}
\alias{sort_table_sqlite}
\title{Sort table from an specified SQLite database}
\usage{
sort_table_sqlite(sqlite_name, table_name, sql_sort_clause)
}
\arguments{
\item{sqlite_name}{A character string specifying the unquoted SQLite
database name}

\item{table_name}{A character string specifying the unquoted SQLite
table name}

\item{sql_sort_clause}{A character string containing SQL order by clause}
}
\description{
Sort table from an specified SQLite database in an specific
path (/R Output/SQLite Files/)
}
