% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite_interface.R
\name{write_sqlite}
\alias{write_sqlite}
\title{Copy data frames to SQLite database}
\usage{
write_sqlite(sqlite_name, table_name, df, ...)
}
\arguments{
\item{sqlite_name}{A character string specifying the unquoted SQLite
database name}

\item{table_name}{A character string specifying the unquoted SQLite
table name}

\item{df}{A data.frame (or coercible to data.frame)}

\item{...}{Other parameters passed on to
\code{\link[DBI]{dbWriteTable}}.}
}
\description{
Writes, overwrites or appends a data frame to SQLite database in an
specific path (/R Output/SQLite Files/).
}
